Eliminate more duplication between default and collection rows.
